Planets are now rendering in the editor. Yay!
