+++ client/src/components/Starmap/Star/index.tsx
@@ -1,142 +1,88 @@
-import React, {Suspense} from "react";
-import {
-  TextureLoader,
-  RepeatWrapping,
-  Mesh,
-  ShaderMaterial,
-  Color,
-  Vector3,
-  AdditiveBlending,
-  Group,
-  Texture,
-} from "three";
-import LensFlare from "./lensFlare";
-import {fragment, vertex} from "./shaders";
-import getUniforms from "./uniforms";
-import ColorUtil from "color";
-import {useTexture} from "@react-three/drei";
-import {useFrame} from "@react-three/fiber";
+import React from "react";
+import {solarRadiusToKilometers} from "server/src/utils/unitTypes";
+import {Color} from "three";
+import OrbitContainer from "../OrbitContainer";
+import Star from "./StarMesh";
+import Selected from "../Selected";
 import {useStarmapStore} from "../starmapStore";
+import {getOrbitPosition} from "server/src/utils/getOrbitPosition";
+import StarPlugin from "server/src/classes/Plugins/Universe/Star";
 
-import texturePath from "./textures/01_Texture.jpg";
-import spritePath from "./textures/Star.svg";
-const distanceVector = new Vector3();
-
-const StarSprite = ({color1}: {color1?: number | Color}) => {
-  const spriteMap = useTexture(spritePath) as Texture;
-
-  return (
-    <sprite>
-      <spriteMaterial
-        attach="material"
-        map={spriteMap}
-        color={color1}
-        sizeAttenuation={false}
-      ></spriteMaterial>
-    </sprite>
+const StarEntity: React.FC<{
+  star: StarPlugin;
+}> = ({star}) => {
+  // const selectedId = useConfigStore(store => store.selectedObject?.id);
+  const selected = false; //selectedId === entity.id;
+  if (!star.satellite) return null;
+  const {semiMajorAxis, eccentricity, orbitalArc, inclination, showOrbit} =
+    star.satellite;
+  const color1 = new Color(
+    `hsl(${star.hue}, 100%, ${star.isWhite ? 100 : 50}%)`
   );
-};
-const SPRITE_SCALE_FACTOR = 50;
-const Star: React.FC<{
-  color1?: number | Color;
-  color2?: number | Color;
-  size?: number;
-  position?: Vector3 | [number, number, number];
-  noLensFlare?: boolean;
-  showSprite?: boolean;
-}> = ({
-  color1 = 0x224488,
-  color2 = 0xf6fcff,
-  size,
-  noLensFlare,
-  showSprite,
-  ...props
-}) => {
-  const texture = React.useMemo(() => {
-    const loader = new TextureLoader();
-    const texture = loader.load(texturePath);
-    texture.wrapS = RepeatWrapping;
-    texture.wrapT = RepeatWrapping;
-    return texture;
-  }, [texturePath]);
-  const uniforms = React.useMemo(
-    () => getUniforms({map: texture, color1, color2}),
-    [color1, color2, texture]
+  const color2 = new Color(
+    `hsl(${star.hue + 20}, 100%, ${star.isWhite ? 100 : 50}%)`
   );
-  const shader = React.useRef<Mesh>();
-  const starMesh = React.useRef<Group>();
-  const starSprite = React.useRef<Group>();
 
   const viewingMode = useStarmapStore(state => state.viewingMode);
-
-  useFrame(({camera}) => {
-    shader.current?.quaternion.copy(camera.quaternion);
-    if (shader.current) {
-      const mat = shader.current.material as ShaderMaterial;
-      mat.uniforms.time.value += 0.03;
-      mat.uniforms.color1.value = color1;
-      mat.uniforms.color2.value = color2;
-    }
-
-    const distance = camera.position.distanceTo(
-      distanceVector.set(camera.position.x, 0, camera.position.z)
-    );
-    if (starSprite.current && starMesh.current) {
-      if (
-        size &&
-        distance / size > 100 &&
-        (useStarmapStore.getState().viewingMode === "core" || showSprite)
-      ) {
-        starSprite.current.visible = true;
-        starMesh.current.visible = false;
-      } else {
-        starSprite.current.visible = false;
-        starMesh.current.visible = true;
-      }
-    }
-  });
-  const color = React.useMemo(() => {
-    if (typeof color1 === "number") {
-      const color = color1.toString(16);
-      return `#${color}`;
-    }
-    const color = color1.toArray();
-    const colorVal = `rgb(${Math.round(color[0] * 255)},${Math.round(
-      color[1] * 255
-    )},${Math.round(color[2] * 255)})`;
-    return ColorUtil(colorVal).lighten(90).rgbNumber();
-  }, [color1]);
-
-  const spriteScale = 1 / (size || 1) / SPRITE_SCALE_FACTOR;
+  const size =
+    viewingMode === "editor"
+      ? 10 + 5 * star.radius
+      : solarRadiusToKilometers(star.radius);
   return (
-    <group {...props}>
-      <pointLight intensity={0.8} decay={2} color={color} castShadow />
-      <group ref={starSprite} scale={[spriteScale, spriteScale, spriteScale]}>
-        <Suspense fallback={null}>
-          <StarSprite color1={color1} />
-        </Suspense>
+    <OrbitContainer
+      radius={semiMajorAxis}
+      eccentricity={eccentricity}
+      orbitalArc={orbitalArc}
+      orbitalInclination={inclination}
+      showOrbit={showOrbit}
+    >
+      <group
+        onPointerOver={() => {
+          if (viewingMode === "viewscreen") return;
+          if (viewingMode === "core") return;
+          document.body.style.cursor = "pointer";
+          const position = getOrbitPosition({
+            eccentricity,
+            orbitalArc,
+            inclination: inclination,
+            semiMajorAxis: semiMajorAxis,
+          });
+          // useConfigStore.setState({
+          //   hoveredPosition: position,
+          //   scaledHoveredPosition: position,
+          // });
+        }}
+        onPointerOut={() => {
+          if (viewingMode === "viewscreen") return;
+          if (viewingMode === "core") return;
+          document.body.style.cursor = "auto";
+          // useConfigStore.setState({
+          //   hoveredPosition: null,
+          //   scaledHoveredPosition: null,
+          // });
+        }}
+        onClick={() => {
+          if (viewingMode === "viewscreen") return;
+          if (viewingMode === "core") return;
+          const position = getOrbitPosition({
+            eccentricity,
+            orbitalArc,
+            inclination: inclination,
+            semiMajorAxis: semiMajorAxis,
+          });
+          useStarmapStore.setState({
+            selectedObjectId: star.name,
+            // selectedPosition: position,
+            // scaledSelectedPosition: position,
+          });
+        }}
+        scale={[size, size, size]}
+      >
+        {selected && viewingMode !== "viewscreen" && <Selected />}
+        <Star color1={color1} color2={color2} size={size} />
       </group>
-      <group ref={starMesh}>
-        <mesh ref={shader} uuid="My star">
-          <circleBufferGeometry attach="geometry" args={[1, 8, 8]} />
-          <shaderMaterial
-            attach="material"
-            fragmentShader={fragment}
-            vertexShader={vertex}
-            uniforms={uniforms}
-            blending={AdditiveBlending}
-            transparent
-            depthTest={false}
-          />
-        </mesh>
-        <mesh uuid="My star background">
-          <sphereBufferGeometry attach="geometry" args={[0.5, 32, 32]} />
-          <meshBasicMaterial attach="material" color={0x000000} />
-        </mesh>
-      </group>
-      {viewingMode !== "core" && !noLensFlare && <LensFlare />}
-    </group>
+    </OrbitContainer>
   );
 };
 
-export default Star;
+export default StarEntity;
